refactor: organize source files into commands and frameworks directories
Move command modules into `src/commands/` and rename `src/framework/` to `src/frameworks/`. Replace `import.meta.env` usage with a centralized `src/env.ts` module using Valibot validation and `process.env`.
