+++ src/commands/logout.ts
@@ -1,6 +1,6 @@
 import { parseArgs } from "node:util";
 
-import { logout as baseLogout } from "./lib/auth";
+import { logout as baseLogout } from "../lib/auth";
 
 const HELP = `
 Log out of Prismic.
