+++ src/frameworks/nextjs.templates.ts
