+++ src/frameworks/nuxt.templates.ts
