+++ src/lib/sentry.ts
@@ -1,22 +1,19 @@
 import * as Sentry from "@sentry/node-core/light";
 
 import packageJson from "../../package.json" with { type: "json" };
+import { env } from "../env";
 
 const SENTRY_DSN =
-	import.meta.env.PRISMIC_SENTRY_DSN ||
+	env.PRISMIC_SENTRY_DSN ||
 	"https://e1886b1775bd397cd1afc60bfd2ebfc8@o146123.ingest.us.sentry.io/4510445143588864";
 
 function isSentryEnabled(): boolean {
-	if (import.meta.env.PRISMIC_SENTRY_ENABLED === undefined) {
-		return import.meta.env.PROD;
-	}
-
-	return import.meta.env.PRISMIC_SENTRY_ENABLED === "true";
+	return env.PRISMIC_SENTRY_ENABLED ?? env.PROD;
 }
 
 function detectEnvironment(): string {
-	if (import.meta.env.PRISMIC_SENTRY_ENVIRONMENT) {
-		return import.meta.env.PRISMIC_SENTRY_ENVIRONMENT;
+	if (env.PRISMIC_SENTRY_ENVIRONMENT) {
+		return env.PRISMIC_SENTRY_ENVIRONMENT;
 	}
 	const prereleaseMatch = packageJson.version.match(/-(.+?)\./);
 	return prereleaseMatch ? prereleaseMatch[1] : "production";
