+++ tsdown.config.ts
@@ -1,4 +1,3 @@
-import { existsSync } from "node:fs";
 import { defineConfig } from "tsdown";
 
 const MODE = process.env.MODE || "production";
@@ -8,14 +7,10 @@ export default defineConfig({
 	format: "esm",
 	platform: "node",
 	minify: true,
-	envFile: existsSync(".env.local") ? ".env.local" : undefined,
 	envPrefix: "PRISMIC_",
 	define: {
-		"import.meta.env.MODE": JSON.stringify(MODE),
-		"import.meta.env.DEV": JSON.stringify(MODE !== "production"),
-		"import.meta.env.PROD": JSON.stringify(MODE === "production"),
-		"import.meta.env.PRISMIC_SENTRY_DSN": "undefined",
-		"import.meta.env.PRISMIC_SENTRY_ENVIRONMENT": "undefined",
-		"import.meta.env.PRISMIC_SENTRY_ENABLED": "undefined",
+		"process.env.MODE": JSON.stringify(MODE),
+		"process.env.DEV": JSON.stringify(MODE !== "production"),
+		"process.env.PROD": JSON.stringify(MODE === "production"),
 	},
 });
